+++ models/appstate.go
@@ -3,7 +3,7 @@ package models
 type AppState struct {
 	HardwareID            string  `json:"hwid" db:"hwid"`
 	ApiToken              *string `json:"token" db:"token"`
-	CurrentEvent          *int64  `json:"-" db:"current_event"`
+	CurrentEvent          *int    `json:"-" db:"current_event"`
 	LastAppliedMigrations int     `json:"-" db:"last_applied_migration"`
 
 	CurrentEventObj *Event `json:"current_event" db:"-"`
